# fix accuracy of IteratorRandom::choose()