Fix accuracy bug in #1058. This is a value-stability-breaking change.