@dhardy 2020-10-12
Could you add before/after benchmarks please?