@kazcw 2020-10-13
Ran benches.

The commit removing the special case for `upper == Some(lower)` caused a substantial performance regression in seq_iter_choose_from_1000. Upon some investigation, it's because when compiling `<iter::SliceMut as IteratorRandom>::choose`, the optimizer knows that `upper == Some(lower)` after inlining `SliceMut::size_hint`, and separating that case enables the optimizer to in effect specialize the function for an ExactSizeIterator by removing the whole loop. Neat! I'll add a comment in the source about why that's there.

After backing that commit out, the two benchmarks affected by these changes are seq_iter_unhinted_choose_from_1000 and seq_iter_window_hinted_choose_from_1000.

master:
test seq_iter_unhinted_choose_from_1000      ... bench:      17,443 ns/iter (+/- 312)
test seq_iter_window_hinted_choose_from_1000 ... bench:       3,330 ns/iter (+/- 77)

this branch:
test seq_iter_unhinted_choose_from_1000      ... bench:      13,387 ns/iter (+/- 162)
test seq_iter_window_hinted_choose_from_1000 ... bench:       3,091 ns/iter (+/- 105)