@vks 2020-10-15
The value stability tests seem to be failing. Could you please fix this?

> Can you squash-merge?

Why squash? The commits seem to be independent enough.