@dhardy 2020-10-17
> Why squash? The commits seem to be independent enough.

Because the individual commits are tiny, and GitHub preserves the individual commits in the PR anyway (e.g. https://github.com/rust-random/rand/commit/05a7ab3f6884c86593b7dda4988218fcd5888093 is a squash while its PR has 16 commits). For larger PRs I don't think this is optimal but for small ones it seems sensible, if not really significant.